+++ src/core/public/index.ts
@@ -106,9 +106,11 @@ import { KeyboardShortcutSetup, KeyboardShortcutStart } from './keyboard_shortcu
 import { ChatServiceSetup, ChatServiceStart } from './chat';
 import type { TelemetryServiceSetup, TelemetryServiceStart } from './telemetry';
 import { HealthCheckServiceSetup, HealthCheckServiceStart } from './healthcheck';
+import type { WazuhBuildInfo } from '../types/wazuh_build_info';
 
 export type { Logos } from '../common';
 export { PackageInfo, EnvironmentMode } from '../server/types';
+export type { WazuhBuildInfo } from '../types/wazuh_build_info';
 /** @interal */
 export { CoreContext, CoreSystem } from './core_system';
 export {
@@ -288,6 +290,8 @@ export interface CoreSetup<TPluginsStart extends object = object, TStart = unkno
   injectedMetadata: {
     getInjectedVar: (name: string, defaultValue?: any) => unknown;
     getBranding: () => Branding;
+    // Wazuh
+    getWazuhBuildInfo: () => WazuhBuildInfo;
   };
   /** {@link StartServicesAccessor} */
   getStartServices: StartServicesAccessor<TPluginsStart, TStart>;
@@ -355,6 +359,8 @@ export interface CoreStart {
   injectedMetadata: {
     getInjectedVar: (name: string, defaultValue?: any) => unknown;
     getBranding: () => Branding;
+    // Wazuh
+    getWazuhBuildInfo: () => WazuhBuildInfo;
   };
   /** {@link WorkspacesStart} */
   workspaces: WorkspacesStart;